+++ bigquery/CHANGES.md
@@ -10,6 +10,8 @@
         mapred.bq.query.results.table.delete
         mapred.bq.input.sharded.export.enable
 
+1.  Remove obsolete `mapred.bq.output.async.write.enabled` property.
+
 ### 0.13.14 - 2019-02-13
 
 1.  POM updates for GCS connector 1.9.14.